requirements.txt
# Core ML/RL libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
gymnasium>=0.29.0
stable-baselines3>=2.0.0
sb3-contrib>=2.0.0

# Configuration and utilities
omegaconf>=2.3.0
hydra-core>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# NLP and dialogue
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0
nltk>=3.8.0
spacy>=3.6.0

# Evaluation and metrics
scikit-learn>=1.3.0
rouge-score>=0.1.2
sacrebleu>=2.3.0

# Demo and visualization
streamlit>=1.25.0
plotly>=5.15.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
ruff>=0.0.280
mypy>=1.4.0
pre-commit>=3.3.0

# Optional: experiment tracking
wandb>=0.15.0
mlflow>=2.5.0
